+++ inst/examples/PPT.ReplaceTextByGraphicExample.R
@@ -1,15 +1,19 @@
 \dontrun{
 
-# # add graphic to slides with matching and remove text
-# 
-# ppt <- PPT.Open("inst/template.pptx", method="RDCOMClient")
-# 
-# what <- "[[tag 1]]"
-# PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_1.png")
-# 
-# # Note that the text appears twice and the graphic is inserted twice with a warning.
-# what <- "[[tag 2]]"
-# PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_2.png")
+file <- system.file("template.pptx", package = "R2PPTaddons")
+ppt <- PPT.Open(file, method="RDCOMClient")
+
+# add graphic to slides with matching text and remove text. Note that by default
+# only text in text boxes is replaced. Text inside rectangles, for example, is
+# not touched unless explicitly prompted. Note also that additional arguments
+# are passed on to PPT.AddGraphicstoSlide2 to specify the position and size of
+# the shape.
+what <- "[[tag 1]]"
+PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_1.png", width = .6)
+
+# Note that the text appears twice and the graphic is inserted twice with a warning.
+what <- "[[tag 2]]"
+PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_2.png")
 
 }