+++ NEWS.md
@@ -1,9 +1,11 @@
 # NEWS
 
 ## TODO
-* No open points
+* PPT.ReplaceShapeByGraphic: Find shape using text pattern and replace by image
  
-## 0.3 (under preperation)
+## 0.3 (under development)
+
+* PPT.ReplaceGraphic: the shapes to operate on (e.g. only text fields) can now be specified.
 * PPT.AddGraphicstoSlideExample rewritten, not compatible with older version any more
 * PPT.FitGraphicInShape2 function to place graphics inside shape
 * Add frame of reference to PPT.AddGraphicstoSlide2 to allow better positioning